+++ shaders/_prelude.vert.glsl
@@ -1,9 +1,5 @@
 #version 300 es
-
-// Only used on older version
-//#define in attribute
-//#define out varying
-precision highp float;
+precision mediump float;
 
 in float feature;
 out float v_feature;
